requirements/base.txt
Django==1.7
argparse==1.2.1
certifi==14.05.14
django-crispy-forms==1.4.0
django-json-rpc==0.6.1
django-tables2==0.15.0
jsonrpc==1.2
lxml==3.3.6
requests==2.4.0
six==1.7.3
wsgiref==0.1.2
